refactor setups
