+++ bin/lib/vets-api/setups/base.rb
@@ -59,29 +59,23 @@ module VetsApi
       end
 
       def set_local_betamocks_cache_dir
-        existing_settings = YAML.safe_load(File.read('config/settings.local.yml'), permitted_classes: [Symbol])
         cache_settings = { 'betamocks' => { 'cache_dir' => '../vets-api-mockdata' } }
         if existing_settings.keys.include?('betamocks')
           puts 'Skipping betamocks cache_dir setting (setting already exists)'
         else
           print 'Editing config/settings.local.yml to set cache_dir for betamocks ...'
-          File.open('config/settings.local.yml', 'a') do |file|
-            file.puts cache_settings.to_yaml.tr('---', '')
-          end
+          save_settings(cache_settings)
           puts 'Done'
         end
       end
 
       def disable_signed_authentication_requests
-        existing_settings = YAML.safe_load(File.read('config/settings.local.yml'), permitted_classes: [Symbol])
         saml_settings = { 'saml' => { 'authn_requests_signed' => false } }
         if existing_settings.keys.include?('saml')
           puts 'Skipping disable signed authentication request (setting already exists)'
         else
           print 'Editing config/settings.local.yml to disable signed authentication requests...'
-          File.open('config/settings.local.yml', 'a') do |file|
-            file.puts saml_settings.to_yaml.tr('---', '')
-          end
+          save_settings(saml_settings)
           puts 'Done'
         end
       end
@@ -108,6 +102,16 @@ module VetsApi
           puts
           puts "\e[1;4m**Please do not commit a Gemfile.lock that does not include sidekiq-pro and sidekiq-ent**\e[0m"
           puts
+        end
+      end
+
+      def existing_settings
+        existing_settings = YAML.safe_load(File.read('config/settings.local.yml'), permitted_classes: [Symbol])
+      end
+
+      def save_settings(settings)
+        File.open('config/settings.local.yml', 'a') do |file|
+          file.puts settings.to_yaml.tr('---', '')
         end
       end
     end
